Use roslaunch arg resolution in rosparam text
Would be very useful to do things like

{{{
rosparam param="app_lists">$(arg app_lists)</rosparam>
}}}

or even use `$(find ...)` embedded in the text. This works for `<param.../>` where the value is a tag attribute. 
Obvious advantage with rosparam is to be able to parse lists and dictionaries.

The fix here is a simple one-liner. It works, I'm pretty sure there are no pitfalls, but it could use a check from someone with good knowledge of the python infra.