Fix test_roslaunch_deps_main
(Unrelated to substitution nesting)

Commit 82f6fed15ba6574be56dbb6f543f5316840008c3 moved test files from roslaunch
to a new test_roslaunch package, but didn't update the path to the "rosmaster.test"
file in the test_roslaunch_deps_main().

The test was still passing, but it wasn't testing what it was supposed
to (ie. that it fails with files from a different package; it was just
throwing non-existent file).