Add roslz4 package
Contains C++ and python libraries for v1.4 of the lz4 streaming format.
Both libraries use the system version of lz4 and use default values for
the stream parameters (e.g., includes a stream checksum, but not block
checksums).  See:
http://fastcompression.blogspot.fr/2013/04/lz4-streaming-format-final.html

The C++ library provide a stream style API for
compressing/decompressing data, similar to the bzip library.  They also
contain convenience buffer to buffer compression/decompression.

The python library provides an API similar to the bz2 module and works
with python 2 and 3.