Avoid adding a slash if left is empty.
If the string 'left' is empty append() inserts a slash in front of 'right', which changes it to a global name. This behaviour is neither documented nor intuitive.
Returning 'right' unchanged in case 'left' is empty removes the issue.
The change may effect existing code, but only if someone knew the behaviour and used it intentionally (IMHO). Otherwise it leads to expected behaviour.