Alternative versions of ros::param::param and ros::NodeHandle::param with return value
Template methods ros::param::param() and ros::NodeHandle::param() set a
variable either to the value of a parameter or, if not available, a predefined
default value. Often the variable is passed straight to a constructor or
function call and then disregarded.

The alternative versions of the param() methods introduced here enable
retrieving the result value from a function return, rather than through an
output variable. This makes possible to write more streamlined code.

For example, the code below:

    std::string path;
    std::string format;
    double fps;

    param("~path", path, "video.mpg");
    param("~format", format, "MPEG");
    param("~fps", fps, 30.0);

    VideoRecorder recorder(path, format, fps);

Could be more succinctly written as:

    VideoRecorder recorder(
        param<std::string>("~path", "video.mpg"),
        param<std::string>("~format", "MPEG"),
        param<double>("~fps", 30.0)
    );

Or:

    NodeHandle nh("~");

    VideoRecorder recorder(
        nh.param<std::string>("path", "video.mpg"),
        nh.param<std::string>("format", "MPEG"),
        nh.param<double>("fps", 30.0)
    );

Signed-off-by: Helio Perroni Filho <xperroni@gmail.com>