Adds support for loading parameters from stdin as well as dumping rosparam to stdout.
If no filename is passed to rosparam load or if '-' is passed as
the filename, rosparam will attempt to read yaml from stdin. This
allows piping yaml into rosparam.

If no filename is specified to rosparam dump or if the filename is
"-", dumps the rosparam yaml to stdout, allowing it to be piped
into other programs.