Add `-w` and `-t` options to roscore.
These define the number of worker-threads of the master and the socket
timeout to use.

For large projects, the default of 3 workers with the OS' default
timeout (of roughly 2-3mins) will cause the topic notification queue to
stall as soon as a computer doesn't answer. So large projects want to
set the timeout much lower and increase the thread-pool size a bit.