Shutdown rospy timers cleanly
Timers go off and start periodically running in the background, and will throw exceptions when ros shutdown occurs because of the sleep function.

These are impossible to catch as they're off in a background thread and can be cleanly handled as in the code change.

Others with the same problem:

* [baxter examples](https://github.com/RethinkRobotics/baxter_examples/issues/48)