Enabled tests
Fixed most tests by replacing test_ros with test_rosmaster, which is apparently
the new location for the messages and services that this test suite needs. There
are still a few failing tests, which look like actual problems, related to how
rosmsg, in raw mode, is resolving types that appear elsewhere, like `Empty`.