roslaunch: Fix <param command="..." /> for python3.
In python3, the standard output of the subprocess is of type bytes. That
causes the data to be stored as a binary xmlrpc node which ends up
base64 encoded when retrieving it later. Additionally, it crashed the
rosparam command line utility which could not deal with that.

This commit checks if the read command output is not of type `str` and
decodes it as UTF-8 in that case. This is a NOP in python2. We can't
check if the read output is of type bytes, because in python2 `bytes`
is just an alias for `str`.