Added the "export_all_args" attribute to the include tag.
This enables a parent launch file to automatically export all args defined in an included launch file. This means that any arg defined in the included launch file will be automatically passed to the include tag, and also added as args to the parent launch file. Additionally, any documentation strings set for the arg will be exported as well.

This eliminates the need for args to be added to each file in a chain of launch files in order to pass an argument down the chain.