Add more logging to publisher update calls.
Because the publisher update calls are critical for setting
up topic connections between subscribers and publishers,
have added more detailed logging messages.  The publishers
are included so you know what the subscriber was told, as
well as when the call succeeded or failed and if so why.
The rosout topic is filtered out to reduce noise.