Add message generation 4 to test_rosbag
This allows to check for the following case:

Consider the following scenario for a message G with the following
versions:

G0 -- rule.bmr --> G1 -- auto-generated rule --> G3
where G3 is the current msg, and G2 is the bag message that needs
migration.

We need to check for a migration rule like this first:

G2 --> G3
instead of the backwards ones that were generated before this bug fix:

G2 --> G1

Note that the lazy check to mark the rules as valid in the
make_update_rule (see
https://github.com/clearpathrobotics/ros_comm/blob/kinetic-devel-cpr/tools/rosbag/src/rosbag/migration.py#L1123)
method makes that rule pass initially, but later it fails because
there's no rule for the sub fields.

Here, in test_rosbag, the messages that corresponds to G in the example
above, is MessageMixed.