Force a write of all user-space buffered data for stdout in Formatter::print (#1012)
* Force a write of all user-space buffered data for stdout in Formatter::print

When application output is connected to pipe a fully-buffered output is used by
default. In this case the logs of application will not be visible for user
until there are enough logs to fill the buffer.

An example where such behaviour can be observed is launching binary ROS executable
from within Android APK. There is not way to connect stdout to TTY and only
pipe can be used. As a result the user cannot see the application output.

The proposed solution is to force flushing of stdout on every line in ROS logger
when a special environment variable is set.

* minor style change