add time when timer expired to timer events
When the a timer expires, the callback is added to the queue.
So the callback is then then poped from the queue at some later point (for which the time is recored as current_real in the timer event).
This adds the time of when the timer actually expired to the timer event info,
in the hope that this will make debugging easier and provide a better base to test the timers.