Fixed corrupted messages when reopening a rosbag with a different file
* Bag::decompressed_chunk_ was not reset in close(). decompressed_chunk_
stores the offset (within the file) of the compressed chunk whose
decompressed content is currently stored in decompress_buffer_. Since it
wasn't reset in close(), if the offset is the same as that of the first
chunk in the file opened next, Bag::decompressChunk() assumed that
it had already decompressed that chunk in the new file and just returned.
As a result, some member functions of Bag worked on the decompressed
chunk of the old bag.
* Cleaned up more state in Bag::close() and ChunkedFile::close()