Specialize BagCallbackT for MessageInstance (#1374)
Call the callback directly with the MessageInstance, so the client
callback can access the connection information.
This is useful when processing a bag file to create a new one b/c we can
keep the message time and connection information, e.g. the latch
property. Also added BagPlayer unit test.