roslaunch: Add ability to read from param server
Add two new roslaunch substitution args, param and optparam.
These function similarily to env and optenv, but read from what is in
the param server at the time of argument substitution. This means the
params for these substitutions are read *prior* to whatever ros
params are set in the launch file. This is useful for storing global
configuration params in the server by some prior launch file or
process prior to launching various nodes that depend on those
configurations. This mechanism is very useful to alter launch file
behavior based on sensor configuration or other global configuration,
like sim time being enabled.