rospy: add recv timeout on header
This recv is done on the thread that handles accepting new connections
to this node. If for some reason the client fails to send the header for
an extended period of time it will prevent any other connections from
being made to this node. If a client hasn't sent the ros header within
the timeout we will drop there connection and move on. In most cases the
client will retry the connection.