rospy: fix tcp reconnect case
If we hit one of the cases where we get an exception during the
connection sequence and a reconnection is desired then we need to close
the socket and set it to None. This will allow the connect loop in
robust_connect_subscriber() to continue for another iteration. Without
these changes the connect loop would exit because conn.socket was not
None and then enter the receive loop without successfully having
negotiated a connection.

This can cause a deadlock between a publisher and subscriber if the
connection was established but the ros header was not written out. This
will leave the publisher trying to receive the ros header from the
subscriber and the subscriber waiting to receive a message from the
publisher. When this case is hit, for whatever reason, it's best to just
close the current connection and establish a new connection.