rosmsg: clean up test dependencies
All dependencies must be declared in the package.xml, and all these
dependencies must be in the ROS distro. rosmsg doesn't meet either of
these requirements today, by using but not depending on both rostest and
std_srvs, and by using test_rosmaster, which isn't in the ROS distro at
all.

Update rosmsg to properly depend upon what it requires, and stop using
test_rosmaster completely. In place of test_rosmaster, use
diagnostic_msgs, which satisfies the same requirements of the existing
tests by having both messages and services as well as standalone and
recursive messages.

Signed-off-by: Kyle Fazzari <kyle@canonical.com>