Fix for #152 add a an optional timeout on service calls
A timeout global timeout is defined for XmlRpc connections.
Overrides are created to allow users to specify a timeout for specific calls.
In particular for ros::master::check a new version is created to allow users to check if the master is up without forever blocking their app.
This fix preserves current functionality but allows user to switch on a global timeout by setting it in the XmlRpcClient class.