fix error in base64EncodedSize()
The base64 encoder called by XmlRpcValue.cpp may write two newlines to the
end of the encoded string, but this is not accounted for in base64EncodedSize().
This results in a buffer overflow, which then causes ROS UDP Transport
connection setup to fail.

One case which results in the double newline is when the total length
of "node_name + topic_name + msg_type" is 33 bytes.

Signed-off-by: Mike Scheutzow <mjs973@hotmail.com>