rospy: Fix severe bug in get_param_cached()
This bug is severe because it can break ordinary rospy.get_param()
after a rospy.get_param_cached() has been perfomed. And, furthermore,
rospy.log*() does a rospy.get_param_cached()!

When a parameter is set in the cache, the cached value will always be
returned, even when the caller uses ordinary rospy.get_param(). This is
as designed, as subscribeParam() is done to make the cache
receive updates from the param server whenever the value changes.

However, existing code only uses the presence of a key in the cache to
indicate whether the cached param is valid. However, this is not enough.

Suppose the param server has these contents:
    top:
        min: 1
        max: 10

    print(rospy.get_param('/top'))
    # So far, so good. Output is:
    # {'min': 1, 'max': 10}

    print(rospy.get_param_cached('/top/min'))
    # As expected, this outputs:
    # 1
    # However, ParamServerCache now contains:
    # {'top': {'min': 1}}
    # This is correct, because only the value of 'min' is cached.

    print(rospy.get_param('/top'))
    # Incorrect output:
    # {'min': 1}
    # Ugh! 'max': 10 is missing!
    # ...because it looks like /top is in the cache, but it was never
    # actually cached; only the individual item within it, 'min', was
    # cached.

The fix is to track the actual keys that are cached in a Python set().
Only return a cached value if either the key or a parent of it is in
the set.

This was found by production code which attempts to get all params using
rospy.get_param('/'). This works OK as long as no calls to
rospy.get_param_cached() have been made. However, the rospy.log*()
functions all do
rospy.get_param_cached('rosout_disable_topics_generation'). So after the
first loginfo, subsequent calls to rospy.get_param('/') return only the
single cached key rather than all params.

Fixes: d16296acb755 ("rospy: added get_param_cached (#1515)")