rospy: Fix performance get_param_cached('non_exist')
If rospy.get_param_cached() is called with a non-existent parameter,
existing code would attempt to encode that the parameter doesn't exist
by storing its value in the cache as an empty dictionary. And it would
correctly make a subscribeParam call to rosmaster to receive updates.

However, every subsequent .get_param_cached()
on this non-existent parameter would re-subscribe for updates from
rosmaster, nullifying any performance benefit from caching. Likewise, a
regular .get_param() would simply re-read the parameter from rosmaster
every time, still not benefitting from caching.

To make matters worse, the rospy.log* functions do a
rospy.get_param_cached('rosout_disable_topics_generation'), so we pay
this penalty on every log message whenever that paremeter is not set (it
is not set by default!).

The cause of this problem is that the translation of an empty dict in
the cache to a KeyError occurs within the cache itself, so the
caller can't know whether the parameter is simply unset vs. not
subscribed for updates.

The fix is simple: Move the check for empty dict from the cache itself
to the caller.

Fixes: d16296acb755 ("rospy: added get_param_cached (#1515)")