Rosbag Recorder: Fix latched topic bug
This bug affected latched topics that were published more than once
per recording session, and only when the `--repeat_latched` option
was set.

Latched messages were added to a std::map object using the `insert`
call. This call adds a new element to a map only if one does not yet
exist for that key. This causes the first latched messages to be the
only ones recorded at the start of each split bagfile, not the most
recent messages.