Change the default tcp_nodelay value from "false" to "true"
Because our ROS system is entirely contained to a local network, we do
not need tcp_nodelay to be "false".

Also worth mentioning, since we are using upstream code for some of our
nodes, we don't have the option to go change the flags passed to all the
subscribers. Instead we change the default here.