roslaunch: Return non-zero exit code for required nodes
If a "required" node in a launch file goes down, we really want to pass
the error code up through the system; specifically to systemd. We can
use this information to do things like attempt restarts. This was an
overlooked feature of roslaunch, which always returns zero if nodes die.