Fix bookkeeping of ongoing calls for recursive calls to CallbackQueue::callAvailable()
When copying over pending calls to the thread-local callback queue held in tls->callbacks, the calling_
member was not updated correctly with the number of added callbacks if tls->callbacks was not empty before.
This may be the case if callAvailable() is called recursively, e.g. from within another callback.

The issue should only affect the empty() and isEmpty() member functions, which apparently are unused
within the CallbackQueue class or within roscpp itself, but may be used by external code.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>